Match Emacs / semantics: integer division, float-zero ±inf
Three corrections, all to keep `/` aligned with Emacs Lisp:

- `Number::Div` of `Int(l)` / `Int(r)` now returns `Int(l / r)`, not
  `Float(l as f64 / r as f64)`. So `(/ 10 3) => 3`, `(/ 7 2 2) => 1`,
  `(integerp (/ 10 3)) => t`. Float promotion only applies when at
  least one operand is `Float`.

- VM `BinaryOp::Div` errored whenever the divisor was `Int(0)`,
  even with a `Float` dividend, so `(/ 1.0 0)` raised
  `OutOfRange: Division by zero`. Tighten the check to `(Int(_),
  Int(0))` and let `Number::Div` produce ±inf for any float
  zero — matching Emacs' `1.0e+INF`.

- The rest-arg defun `/` checked `n == 0`, which `PartialEq<i64>`
  on `Number` matches for both `Int(0)` and `Float(0.0)`, so
  `(funcall '/ 1.0 0.0)` errored where the inline VM path returned
  inf — same call shape, different result. Switch the defun to the
  same `(Int(_), Int(0))` predicate so VM and defun agree.

  While there, fix the single-arg case: `(/ X)` is `1/X` in Emacs
  (`(/ 10) => 0`, `(/ 10.0) => 0.1`); the previous defun returned
  the operand unchanged.

+++ src/number.rs
@@ -166,9 +166,13 @@ impl std::ops::Mul<f64> for Number {
 impl std::ops::Div for Number {
     type Output = Number;
 
+    /// Match Emacs Lisp `/`: integer division when both operands are
+    /// integers, float division otherwise. Callers must guard against
+    /// `Int(0)` divisor before calling — Rust's `i64::div` panics for
+    /// that, but `f64::div` returns `inf`/`nan` (which Emacs surfaces).
     fn div(self, rhs: Self) -> Self::Output {
         match (self, rhs) {
-            (Number::Int(l), Number::Int(r)) => Number::Float(l as f64 / r as f64),
+            (Number::Int(l), Number::Int(r)) => Number::Int(l / r),
             (Number::Int(l), Number::Float(r)) => Number::Float(l as f64 / r),
             (Number::Float(l), Number::Int(r)) => Number::Float(l / r as f64),
             (Number::Float(l), Number::Float(r)) => Number::Float(l / r),
@@ -181,7 +185,7 @@ impl std::ops::Div<i64> for Number {
 
     fn div(self, rhs: i64) -> Self::Output {
         match self {
-            Number::Int(l) => Number::Float(l as f64 / rhs as f64),
+            Number::Int(l) => Number::Int(l / rhs),
             Number::Float(l) => Number::Float(l / rhs as f64),
         }
     }
